requirements.txt
streamlit>=1.20
transformers>=4.30
torch>=1.13
sounddevice>=0.4.6
soundfile>=0.12.1
openai-whisper   # optional; pip install -e may be needed
SpeechRecognition>=3.8.1
pydub>=0.25.1
transitions>=0.9.0
graphviz>=0.20
nltk>=3.8.1
git+https://github.com/aman2656/text2emotion.git
vaderSentiment>=3.3.2
numpy>=1.24
emoji==1.6.3
nltk>=3.8.1